move customizers to seed package and rename files
