+++ consensus/hotstuff/committees/leader/cluster.go
@@ -3,7 +3,6 @@ package leader
 import (
 	"fmt"
 
-	"github.com/onflow/flow-go/model/indices"
 	"github.com/onflow/flow-go/state/protocol"
 	"github.com/onflow/flow-go/state/protocol/seed"
 )
@@ -28,7 +27,7 @@ func SelectionForCluster(cluster protocol.Cluster, epoch protocol.Epoch) (*Leade
 		return nil, fmt.Errorf("could not get leader selection seed for cluster (index: %v) at epoch: %v: %w", cluster.Index(), counter, err)
 	}
 	// create random number generator from the seed and customizer
-	rng, err := seed.PRGFromRandomSource(randomSeed, indices.ProtocolCollectorClusterLeaderSelection(cluster.Index()))
+	rng, err := seed.PRGFromRandomSource(randomSeed, seed.ProtocolCollectorClusterLeaderSelection(cluster.Index()))
 	if err != nil {
 		return nil, fmt.Errorf("could not create rng: %w", err)
 	}
