+++ module/chunks/chunk_assigner.go
@@ -13,6 +13,7 @@ import (
 	"github.com/onflow/flow-go/module/mempool"
 	"github.com/onflow/flow-go/module/mempool/stdmap"
 	"github.com/onflow/flow-go/state/protocol"
+	"github.com/onflow/flow-go/state/protocol/seed"
 )
 
 // DefaultChunkAssignmentAlpha is the default number of verifiers that should be
@@ -97,12 +98,12 @@ func (p *ChunkAssigner) Assign(result *flow.ExecutionResult, blockID flow.Identi
 
 func (p *ChunkAssigner) rngByBlockID(stateSnapshot protocol.Snapshot) (random.Rand, error) {
 	// TODO: seed could be cached to optimize performance
-	seed, err := stateSnapshot.Seed() // potentially returns NoValidChildBlockError
+	randomSource, err := stateSnapshot.Seed() // potentially returns NoValidChildBlockError
 	if err != nil {
 		return nil, fmt.Errorf("failed to retrieve source of randomness: %w", err)
 	}
 
-	rng, err := seed.PRGFromRandomSource(seed, indices.ProtocolVerificationChunkAssignment)
+	rng, err := seed.PRGFromRandomSource(randomSource, seed.ProtocolVerificationChunkAssignment)
 	if err != nil {
 		return nil, fmt.Errorf("failed to instantiate random number generator: %w", err)
 	}
